Move TechnicalCSV into utils.py
And move parse_ob too.
